+++ modules/apps/29-fee/client/cli/query.go
@@ -6,11 +6,99 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/version"
 	"github.com/cosmos/ibc-go/v3/modules/apps/29-fee/types"
 	channeltypes "github.com/cosmos/ibc-go/v3/modules/core/04-channel/types"
 	"github.com/spf13/cobra"
 )
+
+// GetCmdIncentivizedPacket returns the unrelayed incentivized packet for a given packetID
+func GetCmdIncentivizedPacket() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "packet [port-id] [channel-id] [sequence]",
+		Short:   "Query for an unrelayed incentivized packet by port-id, channel-id and packet sequence.",
+		Long:    "Query for an unrelayed incentivized packet by port-id, channel-id and packet sequence.",
+		Args:    cobra.ExactArgs(3),
+		Example: fmt.Sprintf("%s query ibc-fee packet-by-id", version.AppName),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			seq, err := strconv.ParseUint(args[2], 10, 64)
+			if err != nil {
+				return err
+			}
+
+			packetID := channeltypes.PacketId{
+				PortId:    args[0],
+				ChannelId: args[1],
+				Sequence:  seq,
+			}
+
+			req := &types.QueryIncentivizedPacketRequest{
+				PacketId:    packetID,
+				QueryHeight: uint64(clientCtx.Height),
+			}
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			res, err := queryClient.IncentivizedPacket(cmd.Context(), req)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// GetCmdIncentivizedPackets returns all of the unrelayed incentivized packets
+func GetCmdIncentivizedPackets() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "packets",
+		Short:   "Query for all of the unrelayed incentivized packets and associated fees across all channels.",
+		Long:    "Query for all of the unrelayed incentivized packets and associated fees across all channels.",
+		Args:    cobra.NoArgs,
+		Example: fmt.Sprintf("%s query ibc-fee packets", version.AppName),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			pageReq, err := client.ReadPageRequest(cmd.Flags())
+			if err != nil {
+				return err
+			}
+
+			req := &types.QueryIncentivizedPacketsRequest{
+				Pagination:  pageReq,
+				QueryHeight: uint64(clientCtx.Height),
+			}
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			res, err := queryClient.IncentivizedPackets(cmd.Context(), req)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+	flags.AddPaginationFlagsToCmd(cmd, "packets")
+
+	return cmd
+}
 
 // GetCmdTotalRecvFees returns the command handler for the Query/TotalRecvFees rpc.
 func GetCmdTotalRecvFees() *cobra.Command {
@@ -146,6 +234,43 @@ func GetCmdTotalTimeoutFees() *cobra.Command {
 	}
 
 	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// GetCmdCounterpartyAddress returns the command handler for the Query/CounterpartyAddress rpc.
+func GetCmdCounterpartyAddress() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "counterparty-address [channel-id] [address]",
+		Short:   "Query the relayer counterparty address on a given channel",
+		Long:    "Query the relayer counterparty address on a given channel",
+		Args:    cobra.ExactArgs(2),
+		Example: fmt.Sprintf("%s query ibc-fee counterparty-address channel-5 cosmos1layxcsmyye0dc0har9sdfzwckaz8sjwlfsj8zs", version.AppName),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			if _, err := sdk.AccAddressFromBech32(args[1]); err != nil {
+				return err
+			}
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			req := &types.QueryCounterpartyAddressRequest{
+				ChannelId:      args[0],
+				RelayerAddress: args[1],
+			}
+
+			res, err := queryClient.CounterpartyAddress(cmd.Context(), req)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
 
 	return cmd
 }